Customer_Churn_Prediction/frontend/requirements.txt
streamlit==1.55.0
requests==2.32.5
pandas==2.3.3
python-dotenv